+++ lib/aspera/cli/plugins/faspex5.rb
@@ -79,12 +79,14 @@ module Aspera
           )
           return {
             preset_value: {
-              url:           app_url,
-              username:      wiz_username,
-              auth:          :jwt.to_s,
-              private_key:   "@file:#{private_key_path}",
-              client_id:     client_id,
-              client_secret: client_secret
+              url:         app_url,
+              username:    wiz_username,
+              auth:        :jwt.to_s,
+              private_key: "@file:#{private_key_path}",
+              params:      {
+                client_id:     client_id,
+                client_secret: client_secret
+              }
             },
             test_args:    'user profile show'
           }
@@ -198,12 +200,12 @@ module Aspera
             skip_ids_persistency = PersistencyActionOnce.new(
               manager: persistency,
               data:    [],
-              id:      IdGenerator.from_list([
+              id:      IdGenerator.from_list(
                 'faspex_recv',
                 options.get_option(:url, mandatory: true),
                 options.get_option(:username, mandatory: true),
                 options.get_option(:box, mandatory: true)
-              ])
+              )
             )
           end
           packages = []